Inline CTA logic from helper to cell
