+++ decidim-core/app/cells/decidim/content_blocks/hero_cell.rb
@@ -3,9 +3,6 @@
 module Decidim
   module ContentBlocks
     class HeroCell < Decidim::ViewModel
-      include Decidim::CtaButtonHelper
-
-      # Needed so that the `CtaButtonHelper` can work.
       def decidim_participatory_processes
         Decidim::ParticipatoryProcesses::Engine.routes.url_helpers
       end
@@ -31,6 +28,28 @@ module Decidim
         hash << background_image
 
         hash.join(Decidim.cache_key_separator)
+      end
+
+      def cta_button_text
+        translated_attribute(model.settings.cta_button_text).presence || t("decidim.pages.home.hero.participate")
+      end
+
+      def cta_button
+        link_to cta_button_text, cta_button_path, class: "hero-cta button expanded large button--sc", title: t("decidim.pages.home.hero.participate_title")
+      end
+
+      def cta_button_path
+        if model.settings.cta_button_path.present?
+          translated_attribute(model.settings.cta_button_path)
+        elsif Decidim::ParticipatoryProcess.where(organization: current_organization).published.any?
+          decidim_participatory_processes.participatory_processes_path
+        elsif current_user
+          decidim.account_path
+        elsif current_organization.sign_up_enabled?
+          decidim.new_user_registration_path
+        else
+          decidim.new_user_session_path
+        end
       end
     end
   end
