+++ decidim-verifications/spec/views/decidim/verifications/authorizations/new.html.erb_spec.rb
@@ -14,7 +14,6 @@ module Decidim
     before do
       view.extend AuthorizationFormHelper
       view.extend DecidimFormHelper
-      view.extend CtaButtonHelper
 
       allow(view).to receive(:current_organization).and_return(organization)
       allow(view).to receive(:handler).and_return(handler)
